Extract GPUI API primitives into a new gpui_types crate
Introduce `crates/gpui_types`
holding the self-contained value types shared by GPUI and its platform backends:

geometry (`Pixels`, `Point`, `Size`, `Bounds`, ...),
colors (`Rgba`, `Hsla`, `Background`, ...),
and the taffy layout bridge (`AvailableSpace`, ...).

`gpui` now depends on and re-exports `gpui_types`,
so `gpui::Pixels` and `gpui_types::Pixels` name the same type
and downstream code is unchanged.

`gpui_linux` imports its geometry primitives from `gpui_types` directly,
demonstrating that a platform backend can depend on the primitives
without going through `gpui`.

Methods that require `App`/`Window` stay in `gpui`
as the `BoundsExt` and `HasRemSize` extension traits.

Conversions that bridge two foreign types (taffy, lyon) moved to `gpui_types`
to satisfy the orphan rule, with the lyon conversions behind an optional `lyon` feature.
`Pixels.0` and `Background`'s internals became `pub`
so they remain usable across the new boundary.

Release Notes:

- N/A
