+++ Cargo.toml
@@ -97,6 +97,7 @@ members = [
     "crates/gpui_platform",
     "crates/gpui_shared_string",
     "crates/gpui_tokio",
+    "crates/gpui_types",
     "crates/gpui_util",
     "crates/gpui_web",
     "crates/gpui_wgpu",
@@ -365,6 +366,7 @@ gpui_macros = { path = "crates/gpui_macros" }
 gpui_platform = { path = "crates/gpui_platform", default-features = false }
 gpui_shared_string = { path = "crates/gpui_shared_string" }
 gpui_tokio = { path = "crates/gpui_tokio" }
+gpui_types = { path = "crates/gpui_types" }
 gpui_util = { path = "crates/gpui_util" }
 gpui_web = { path = "crates/gpui_web" }
 gpui_wgpu = { path = "crates/gpui_wgpu" }
