+++ crates/gpui/src/taffy.rs
@@ -1,6 +1,6 @@
 use crate::{
-    AbsoluteLength, App, Bounds, DefiniteLength, Edges, GridTemplate, Length, Pixels, Point, Size,
-    Style, Window, size,
+    AbsoluteLength, App, AvailableSpace, Bounds, DefiniteLength, Edges, GridTemplate, Length,
+    Pixels, Point, Size, Style, Window, size,
     util::{
         ceil_to_device_pixel, round_half_toward_zero, round_stroke_to_device_pixel,
         round_to_device_pixel,
@@ -10,9 +10,8 @@ use collections::{FxHashMap, FxHashSet};
 use std::{fmt::Debug, ops::Range};
 use taffy::{
     TaffyTree, TraversePartialTree as _,
-    geometry::{Point as TaffyPoint, Rect as TaffyRect, Size as TaffySize},
+    geometry::{Rect as TaffyRect, Size as TaffySize},
     prelude::{max_content, min_content},
-    style::AvailableSpace as TaffyAvailableSpace,
     tree::NodeId,
 };
 
@@ -597,31 +596,6 @@ impl ToTaffy<taffy::style::Dimension> for AbsoluteLength {
     }
 }
 
-impl<T, T2> From<TaffyPoint<T>> for Point<T2>
-where
-    T: Into<T2>,
-    T2: Clone + Debug + Default + PartialEq,
-{
-    fn from(point: TaffyPoint<T>) -> Point<T2> {
-        Point {
-            x: point.x.into(),
-            y: point.y.into(),
-        }
-    }
-}
-
-impl<T, T2> From<Point<T>> for TaffyPoint<T2>
-where
-    T: Into<T2> + Clone + Debug + Default + PartialEq,
-{
-    fn from(val: Point<T>) -> Self {
-        TaffyPoint {
-            x: val.x.into(),
-            y: val.y.into(),
-        }
-    }
-}
-
 impl<T, U> ToTaffy<TaffySize<U>> for Size<T>
 where
     T: ToTaffy<U> + Clone + Debug + Default + PartialEq,
@@ -644,100 +618,6 @@ where
             right: self.right.to_taffy(rem_size, scale_factor),
             bottom: self.bottom.to_taffy(rem_size, scale_factor),
             left: self.left.to_taffy(rem_size, scale_factor),
-        }
-    }
-}
-
-impl<T, U> From<TaffySize<T>> for Size<U>
-where
-    T: Into<U>,
-    U: Clone + Debug + Default + PartialEq,
-{
-    fn from(taffy_size: TaffySize<T>) -> Self {
-        Size {
-            width: taffy_size.width.into(),
-            height: taffy_size.height.into(),
-        }
-    }
-}
-
-impl<T, U> From<Size<T>> for TaffySize<U>
-where
-    T: Into<U> + Clone + Debug + Default + PartialEq,
-{
-    fn from(size: Size<T>) -> Self {
-        TaffySize {
-            width: size.width.into(),
-            height: size.height.into(),
-        }
-    }
-}
-
-/// The space available for an element to be laid out in
-#[derive(Copy, Clone, Default, Debug, Eq, PartialEq)]
-pub enum AvailableSpace {
-    /// The amount of space available is the specified number of pixels
-    Definite(Pixels),
-    /// The amount of space available is indefinite and the node should be laid out under a min-content constraint
-    #[default]
-    MinContent,
-    /// The amount of space available is indefinite and the node should be laid out under a max-content constraint
-    MaxContent,
-}
-
-impl AvailableSpace {
-    /// Returns a `Size` with both width and height set to `AvailableSpace::MinContent`.
-    ///
-    /// This function is useful when you want to create a `Size` with the minimum content constraints
-    /// for both dimensions.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use gpui::AvailableSpace;
-    /// let min_content_size = AvailableSpace::min_size();
-    /// assert_eq!(min_content_size.width, AvailableSpace::MinContent);
-    /// assert_eq!(min_content_size.height, AvailableSpace::MinContent);
-    /// ```
-    pub const fn min_size() -> Size<Self> {
-        Size {
-            width: Self::MinContent,
-            height: Self::MinContent,
-        }
-    }
-}
-
-impl From<AvailableSpace> for TaffyAvailableSpace {
-    fn from(space: AvailableSpace) -> TaffyAvailableSpace {
-        match space {
-            AvailableSpace::Definite(Pixels(value)) => TaffyAvailableSpace::Definite(value),
-            AvailableSpace::MinContent => TaffyAvailableSpace::MinContent,
-            AvailableSpace::MaxContent => TaffyAvailableSpace::MaxContent,
-        }
-    }
-}
-
-impl From<TaffyAvailableSpace> for AvailableSpace {
-    fn from(space: TaffyAvailableSpace) -> AvailableSpace {
-        match space {
-            TaffyAvailableSpace::Definite(value) => AvailableSpace::Definite(Pixels(value)),
-            TaffyAvailableSpace::MinContent => AvailableSpace::MinContent,
-            TaffyAvailableSpace::MaxContent => AvailableSpace::MaxContent,
-        }
-    }
-}
-
-impl From<Pixels> for AvailableSpace {
-    fn from(pixels: Pixels) -> Self {
-        AvailableSpace::Definite(pixels)
-    }
-}
-
-impl From<Size<Pixels>> for Size<AvailableSpace> {
-    fn from(size: Size<Pixels>) -> Self {
-        Size {
-            width: AvailableSpace::Definite(size.width),
-            height: AvailableSpace::Definite(size.height),
         }
     }
 }
