+++ crates/gpui_linux/Cargo.toml
@@ -58,6 +58,7 @@ bytemuck = "1"
 collections.workspace = true
 futures.workspace = true
 gpui.workspace = true
+gpui_types.workspace = true
 gpui_util.workspace = true
 gpui_wgpu = { workspace = true, optional = true, features = ["font-kit"] }
 http_client.workspace = true
